% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{assert_scalar_character}
\alias{assert_scalar_character}
\title{Assert that an input is a scalar character}
\usage{
assert_scalar_character(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Input to check.}

\item{name}{Variable name for error message.}
}
\description{
Assert that an input is a scalar character
}
