% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile.R
\name{compact_pile}
\alias{compact_pile}
\title{Compact All Runs for a Model}
\usage{
compact_pile(pile, model)
}
\arguments{
\item{pile}{A logpile_pile object.}

\item{model}{Model ID string.}
}
\description{
Compact All Runs for a Model
}
