% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile.R
\name{compact_projections}
\alias{compact_projections}
\title{Compact Projections}
\usage{
compact_projections(pile, id)
}
\arguments{
\item{pile}{A logpile_pile object.}

\item{id}{Projection version string.}
}
\description{
Compact Projections
}
