% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coords}
\alias{coords}
\title{Extract Design Space Coordinates from Pile}
\usage{
coords(model, pile = get_active_pile())
}
\arguments{
\item{model}{The model ID string (e.g. "FF16@v1").}

\item{pile}{A \code{logpile_pile} object.}
}
\value{
A data frame of design coordinates and run statuses.
}
\description{
Scans the index namespace and returns a data frame of evaluated coordinates.
}
