% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile.R
\name{create_pile}
\alias{create_pile}
\title{Create a Logpile Pile Interface}
\usage{
create_pile(path)
}
\arguments{
\item{path}{Absolute or relative path to the pile root directory.}
}
\value{
A \code{logpile_pile} object.
}
\description{
Initializes the storage directory structure and returns
a handle encapsulating the storr index database.
}
