% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{evaluate_request}
\alias{evaluate_request}
\title{Evaluate a Request Idempotently}
\usage{
evaluate_request(req, pile)
}
\arguments{
\item{req}{A request list.}

\item{pile}{A \code{logpile_pile} object.}
}
\value{
A character fingerprint (invisibly).
}
\description{
Checks if a request is already processed, and if not, runs and commits it.
}
