% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{extract_scalars}
\alias{extract_scalars}
\title{Extract length-1 numeric or logical parameters from a list}
\usage{
extract_scalars(lst)
}
\arguments{
\item{lst}{A list of parameters.}
}
\value{
A list of scalar parameters.
}
\description{
Extract length-1 numeric or logical parameters from a list
}
\keyword{internal}
