% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_transform}
\alias{get_transform}
\title{Get transform Function for a Model ID}
\usage{
get_transform(model_id)
}
\arguments{
\item{model_id}{Character model identifier, e.g. "FF16@v1".}
}
\value{
A transform function (or identity if none registered).
}
\description{
Get transform Function for a Model ID
}
