% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{knn}
\alias{knn}
\title{Coverage KNN Query}
\usage{
knn(query, k = 1L, model, pile = get_active_pile())
}
\arguments{
\item{query}{Numeric vector, matrix, or data frame query.}

\item{k}{Number of neighbors to return.}

\item{model}{The model ID string.}

\item{pile}{A \code{logpile_pile} object.}
}
\value{
A list with \code{nn.idx} and \code{nn.dists}.
}
\description{
Performs a nearest-neighbor search of a query point against the existing
design coordinates in the pile.
}
