% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{map_workers}
\alias{map_workers}
\title{Map Workers}
\usage{
map_workers(manifest, pile, workers)
}
\arguments{
\item{manifest}{A manifest object}

\item{pile}{A pile object}

\item{workers}{Integer workers}
}
\description{
Uses crew to distribute requests across multiple R sessions.
}
\keyword{internal}
