% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{merge_and_cast}
\alias{merge_and_cast}
\title{Safely Cast and Merge Input Values with Defaults}
\usage{
merge_and_cast(
  target,
  defaults,
  allowed_extra = character(0),
  context_name = "list"
)
}
\arguments{
\item{target}{Sparse user input list.}

\item{defaults}{Default configuration list.}

\item{allowed_extra}{Vector of valid extra parameter names.}

\item{context_name}{Context string for error reporting.}
}
\value{
A type-safe representation of the parameters.
}
\description{
Safely Cast and Merge Input Values with Defaults
}
\keyword{internal}
