% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{parse_namespace}
\alias{parse_namespace}
\title{Parse Request Fingerprint Namespace}
\usage{
parse_namespace(fp)
}
\arguments{
\item{fp}{Namespace string (e.g. "FF16@v1").}
}
\value{
A list with \code{model} and \code{version} fields.
}
\description{
Parse Request Fingerprint Namespace
}
\keyword{internal}
