% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{place_coords}
\alias{place_coords}
\title{Populate Request Template with Coordinates}
\usage{
place_coords(template, coords)
}
\arguments{
\item{template}{A request template.}

\item{coords}{A list or single-row data frame of coordinates.}
}
\value{
A resolved and populated request list.
}
\description{
Injects parameter values from a coordinate row into the request template.
}
