% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate_set}
\alias{predicate_set}
\title{Define a Predicate Set}
\usage{
predicate_set(ids)
}
\arguments{
\item{ids}{Character vector of registered predicate ids.}
}
\value{
A named list with class \code{predicate_set} and a \code{predicate_set_fingerprint} attribute.
}
\description{
Define a Predicate Set
}
