% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{project_logs}
\alias{project_logs}
\title{Bulk Project Runs}
\usage{
project_logs(fingerprints, proj, model, pile = get_active_pile())
}
\arguments{
\item{fingerprints}{Character vector of run fingerprints.}

\item{proj}{Projection id string or function.}

\item{model}{Model ID string (e.g. "FF16@v1").}

\item{pile}{Pile object.}
}
\value{
A combined data frame.
}
\description{
Evaluates a projection across multiple runs simultaneously using Arrow
dataset pushdown for column selection and row filtering.
}
