% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile.R
\name{read_log}
\alias{read_log}
\title{Read a parquet file, optionally selecting a single run by fingerprint.}
\usage{
read_log(path, row_selector = NULL)
}
\description{
Read a parquet file, optionally selecting a single run by fingerprint.
}
