% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{register_transform}
\alias{register_transform}
\title{Register a transform Function}
\usage{
register_transform(model_prefix, fn)
}
\arguments{
\item{model_prefix}{Character prefix of the model, e.g. "FF16".}

\item{fn}{A function mapping a raw log to an enriched log.}
}
\description{
Register a transform Function
}
