% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile.R
\name{relink_records}
\alias{relink_records}
\title{Point records at a compacted file}
\usage{
relink_records(pile, namespace, fingerprints, rel_path, create = FALSE)
}
\description{
With create = FALSE, records that are absent or not "done" are skipped
(index compaction). With create = TRUE, missing records are minted as
parquet records (bulk projection).
}
