% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{resolve_control}
\alias{resolve_control}
\title{Resolve Control Settings}
\usage{
resolve_control(req_control, schema)
}
\arguments{
\item{req_control}{Request control list.}

\item{schema}{Model schema list.}
}
\value{
A resolved control settings list.
}
\description{
Resolve Control Settings
}
\keyword{internal}
