% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{resolve_drivers}
\alias{resolve_drivers}
\title{Resolve Driver Sets}
\usage{
resolve_drivers(req_drivers)
}
\arguments{
\item{req_drivers}{Request drivers.}
}
\value{
A resolved driver representation (hash string or nested list).
}
\description{
Resolve Driver Sets
}
\keyword{internal}
