% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{resolve_global}
\alias{resolve_global}
\title{Resolve Global Parameters}
\usage{
resolve_global(req_global, schema)
}
\arguments{
\item{req_global}{Request global list.}

\item{schema}{Model schema list.}
}
\value{
A resolved global settings list.
}
\description{
Resolve Global Parameters
}
\keyword{internal}
