% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{resolve_request}
\alias{resolve_request}
\title{Resolve and Validate a Request against the plant Schema}
\usage{
resolve_request(request)
}
\arguments{
\item{request}{A list representing the request.}
}
\value{
A resolved and type-safe request list.
}
\description{
Resolve and Validate a Request against the plant Schema
}
