% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{resolve_schedule}
\alias{resolve_schedule}
\title{Resolve Node Schedule}
\usage{
resolve_schedule(req_schedule, n_strats, max_lifetime)
}
\arguments{
\item{req_schedule}{Request schedule.}

\item{n_strats}{Integer count of strategy parameter maps.}

\item{max_lifetime}{Numeric maximum patch lifetime.}
}
\value{
A resolved node schedule list.
}
\description{
Resolve Node Schedule
}
\keyword{internal}
