% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{size_field}
\alias{size_field}
\title{Create Size-Resolved Field Projection}
\usage{
size_field(B, method = "quantile")
}
\arguments{
\item{B}{Integer number of size classes.}

\item{method}{String, "quantile" or "linear".}
}
\value{
A projection function that resolves cohort state into B size classes.
}
\description{
Create Size-Resolved Field Projection
}
