% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{stand_summary_fn}
\alias{stand_summary_fn}
\title{Stand Summary Projection}
\usage{
stand_summary_fn(df)
}
\arguments{
\item{df}{Transformed log dataframe.}
}
\value{
Aggregated stand summary dataframe.
}
\description{
Generic projection that groups by t and computes a
density-weighted mean for numeric columns, and total density.
}
