% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_ff16}
\alias{transform_ff16}
\title{transform FF16 Log}
\usage{
transform_ff16(log)
}
\arguments{
\item{log}{A raw log data frame.}
}
\value{
Enriched data frame.
}
\description{
Enriches an FF16 raw log with density, basal area, and diameter.
}
